requirements.txt
Django>=4.2,<5
gunicorn
whitenoise>=6.0
dj-database-url
psycopg2-binary
python-dotenv
reportlab

